requirement.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
distlib==0.3.1
Django==3.1.5
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-phonenumber-field==5.0.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
lockfile==0.12.2
mysqlclient==2.0.3
pastel==0.2.1
phonenumbers==8.12.17
promise==2.3
PyJWT==1.7.1
pylev==1.3.0
pytz==2020.5
requests-toolbelt==0.9.1
Rx==1.6.1
shellingham==1.3.2
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tomlkit==0.7.0
virtualenv==20.4.0
